+++ client/src/config/eventSearchTags.ts
@@ -1,3 +1,4 @@
+import { IHalLink } from '../types/IHalLink'
 import { IHalLinks } from '../types/IHalLinks'
 
 /**
@@ -12,4 +13,46 @@ export const relatedObjects: IHalLinks = {
     searchTag: 'lux:eventIncludedItems',
     tab: 'objects',
   },
+}
+
+// Related item and work types
+export const relatedTypes: IHalLinks = {
+  itemTypes: {
+    title: 'Object Types',
+    searchTag: 'lux:eventItemTypes',
+    tab: 'objects',
+    jsonSearchTerm: 'classification',
+  },
+  workTypes: {
+    title: 'Work Types',
+    searchTag: 'lux:eventWorkTypes',
+    tab: 'works',
+    jsonSearchTerm: 'classification',
+  },
+}
+
+// Used for rendering related locations
+export const locations: IHalLink = {
+  title: 'Related Locations',
+  searchTag: 'lux:eventRelatedPlaces',
+  tab: 'places',
+  isSemantic: true,
+}
+
+// Shows related accordions
+export const relatedAccordions: IHalLinks = {
+  subjects: {
+    title: 'Related Subjects',
+    searchTag: 'lux:eventRelatedSubjects',
+    tab: 'objects',
+    jsonSearchTerm: [],
+  },
+  materials: {
+    title: 'Related Materials',
+    searchTag: 'lux:eventRelatedMaterials',
+    tab: 'concepts',
+    jsonSearchTerm: 'material',
+  },
+  ...relatedTypes,
+  locations,
 }