+++ client/src/features/event/EventPage.tsx
@@ -3,7 +3,7 @@ import Col from 'react-bootstrap/Col'
 import Row from 'react-bootstrap/Row'
 import { ErrorBoundary } from 'react-error-boundary'
 
-import { relatedObjects } from '../../config/eventSearchTags'
+import { relatedAccordions, relatedObjects } from '../../config/eventSearchTags'
 import StyledEntityBody from '../../styles/shared/EntityBody'
 import StyledEntityPageSection from '../../styles/shared/EntityPageSection'
 import DataSources from '../common/DataSources'
@@ -13,6 +13,7 @@ import RelatedObjectsAndWorks from '../common/RelatedObjectsAndWorks'
 import { ErrorFallback } from '../error/ErrorFallback'
 import EventParser from '../../lib/parse/data/EventParser'
 import StyledHr from '../../styles/shared/Hr'
+import AccordionContainer from '../relatedLists/AccordionContainer'
 
 import AboutPanel from './AboutPanel'
 import HowDoISeeIt from './HowDoISeeIt'
@@ -39,6 +40,12 @@ const EventPage: React.FC<{ data: any }> = ({ data }) => {
               links={data._links}
               relationships={relatedObjects}
               type="event"
+            />
+          </ErrorBoundary>
+          <ErrorBoundary FallbackComponent={ErrorFallback}>
+            <AccordionContainer
+              searchTags={relatedAccordions}
+              providedHalLinks={event.json._links}
             />
           </ErrorBoundary>
         </Col>